+++ CHANGELOG.md
@@ -1,13 +1,15 @@
 # Change log of DVH Analytics
 
-v0.8.3 (TBD)
+v0.8.3 (2020.06.11)
 --------------------
+ - [Database] Initialize SQL tables if needed on Write Test (avoids crash)
  - [Import] Cap DVH max dose to prevent excessive memory use [Issue 83](https://github.com/cutright/DVH-Analytics/issues/83)
+ - [Import] Save ROI Map in Import Module pushes ROI Map changes to previous imports as well
+ - [Import] When right-clicking ROI, then Add <ROI> as Physician ROI, ensure right-clicked ROI is added as a variation 
+ if a different Physician ROI name is used
  - [Plot Visuals] DVH Line selection and nonselection options available in User Settings
  - [Plot Visuals] Fix for [Issue 86](https://github.com/cutright/DVH-Analytics/issues/86)
  - [ROI Map] ROI Types now stored in ROI Map [Issue 81](https://github.com/cutright/DVH-Analytics/issues/81)
- - [Import] Save ROI Map in Import Module pushes ROI Map changes to previous imports as well
-
 
 v0.8.2 (2020.06.05)
 --------------------